fix date section header
