+++ iOS/UI/Updates/MangaUpdatesView.swift
@@ -75,7 +75,7 @@ struct MangaUpdatesView: View {
                     }
                 }
             } header: {
-                Text(makeRelativeDate(days))
+                Text(makeRelativeDate(days: entries[days].0))
                     .foregroundStyle(.primary)
                     .font(.system(size: 16, weight: .medium))
             }
@@ -84,7 +84,7 @@ struct MangaUpdatesView: View {
 
     var listItems: some View {
         ForEach(entries.indices, id: \.self) { days in
-            Text(makeRelativeDate(days))
+            Text(makeRelativeDate(days: entries[days].0))
                 .foregroundColor(.primary)
                 .font(.system(size: 16, weight: .medium))
 
@@ -152,31 +152,5 @@ struct MangaUpdatesView: View {
         offset = 0
         reachedEnd = false
         await loadNewEntries()
-    }
-
-    private func makeRelativeDate(_ days: Int) -> String {
-        let now = Date()
-        let date = now.addingTimeInterval(-86400 * Double(days))
-        let difference = Calendar.autoupdatingCurrent.dateComponents(Set([Calendar.Component.day]), from: date, to: now)
-
-        // today or yesterday
-        if days < 2 {
-            let formatter = DateFormatter()
-            formatter.locale = Locale.autoupdatingCurrent
-            formatter.dateStyle = .medium
-            formatter.doesRelativeDateFormatting = true
-            return formatter.string(from: date)
-        } else if days < 8 { // n days ago
-            let formatter = DateComponentsFormatter()
-            formatter.unitsStyle = .short
-            formatter.allowedUnits = .day
-            guard let timePhrase = formatter.string(from: difference) else { return "" }
-            return String(format: NSLocalizedString("%@_AGO", comment: ""), timePhrase)
-        } else { // mm/dd/yy
-            let formatter = DateFormatter()
-            formatter.locale = Locale.autoupdatingCurrent
-            formatter.dateStyle = .short
-            return formatter.string(from: date)
-        }
     }
 }
